backend/requirements.txt
Django
djangorestframework
requests
pillow
django-filter